Adds transfromer, validator and controller with tests for offer domain

+++ database/migrations/2018_08_16_150311_create_table_vouchers.php
@@ -19,7 +19,7 @@ class CreateTableVouchers extends Migration
             $table->integer('offer_id')->unsigned();
             $table->integer('recipient_id')->unsigned()->nullable()->default(null);
             $table->dateTime('expires_at')->nullable()->default(null);
-            $table->integer('is_used')->unsigned();
+            $table->tinyInteger('is_used')->unsigned();
             $table->dateTime('used_at')->nullable()->default(0);
             $table->timestamps();
             $table->softDeletes();
